Change structure of index file
